refactor(core): split agent_task_loop_controller god-file into module dir
