+++ homeboy.json
@@ -673,7 +673,7 @@
         "field_patterns::src/commands/trace/output.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_contract.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_fanout.rs::RepeatedFieldPattern",
-        "field_patterns::src/core/agent_task_loop_controller.rs::RepeatedFieldPattern",
+        "field_patterns::src/core/agent_task_loop_controller/policy.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_provider.rs::RepeatedFieldPattern",
         "field_patterns::src/core/agent_task_schedule.rs::RepeatedFieldPattern",
         "field_patterns::src/core/api_jobs/types.rs::RepeatedFieldPattern",
@@ -708,7 +708,7 @@
         "intra-method-duplication::src/commands/fuzz.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/commands/route.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/agent_task_contract.rs::IntraMethodDuplicate",
-        "intra-method-duplication::src/core/agent_task_loop_controller.rs::IntraMethodDuplicate",
+        "intra-method-duplication::src/core/agent_task_loop_controller/record_impl.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/artifact_manifest.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/artifact_origin.rs::IntraMethodDuplicate",
         "intra-method-duplication::src/core/cleanup.rs::IntraMethodDuplicate",
@@ -754,8 +754,6 @@
         "structural::src/core/agent_task_controller_service/tests.rs::GodFile",
         "structural::src/core/agent_task_controller_service/tests.rs::HighItemCount",
         "structural::src/core/agent_task_lifecycle.rs::GodFile",
-        "structural::src/core/agent_task_loop_controller.rs::GodFile",
-        "structural::src/core/agent_task_loop_controller.rs::HighItemCount",
         "structural::src/core/agent_task_promotion.rs::GodFile",
         "structural::src/core/agent_task_promotion.rs::HighItemCount",
         "structural::src/core/agent_task_provider.rs::GodFile",
